feat: refactor form components and add interactive form section with validation
