+++ src/feature/auth/form/components/form-2/FormSection.tsx
@@ -1,73 +1,8 @@
-"use client";
-
-import { useFormStore } from "@/shared/store/useFormStore";
 import Progress from "@/shared/component/Progress";
+import FormSectionInteractive from "@/feature/auth/form/components/form-2/FormSectionInteractive";
 import Kategori from "@/feature/auth/form/components/form-2/Kategori";
-import ButtonNext from "@/shared/component/auth/ButtonNext";
-import PertanyaanSayur from "@/feature/auth/form/components/form-2/PertanyaanSayur";
-import PertanyaanBuah from "@/feature/auth/form/components/form-2/PertanyaanBuah";
-import PertanyaanTernak from "@/feature/auth/form/components/form-2/PertanyaanTernak";
-import PertanyaanDaging from "@/feature/auth/form/components/form-2/PertanyaanDaging";
-import ButtonPrev from "@/shared/component/auth/ButtonPrev";
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { useSupplierDraftForm } from "@/hooks/use-supplier-draft-form";
-import { showFormSectionToast } from "@/shared/component/toast";
-import {
-  buahSchema,
-  dagingSchema,
-  sayurSchema,
-  ternakSchema,
-} from "@/feature/auth/form/types/type";
 
 const FormSection = () => {
-  const { category, answers } = useFormStore();
-  const router = useRouter();
-  const { submitDraftAndFinish, isSubmittingDraft } = useSupplierDraftForm();
-
-  const step = useFormStore((state) => state.step);
-  const setStep = useFormStore((state) => state.setStep);
-  const { isQuestionOpen } = useFormStore();
-
-  useEffect(() => {
-    if (step < 2) {
-      router.push("/form-1");
-    }
-  }, [router, step]);
-
-  const handleSubmit = () => {
-    if (!category) {
-      showFormSectionToast("categoryRequired");
-      return;
-    }
-
-    let result:
-      | ReturnType<typeof dagingSchema.safeParse>
-      | ReturnType<typeof sayurSchema.safeParse>
-      | ReturnType<typeof buahSchema.safeParse>
-      | ReturnType<typeof ternakSchema.safeParse>;
-
-    if (category === "Daging") {
-      result = dagingSchema.safeParse(answers.daging);
-    } else if (category === "Sayur") {
-      result = sayurSchema.safeParse(answers.sayur);
-    } else if (category === "Buah") {
-      result = buahSchema.safeParse(answers.buah);
-    } else {
-      result = ternakSchema.safeParse(answers.ternak);
-    }
-
-    if (!result.success) {
-      showFormSectionToast(
-        "questionRequired",
-        result.error.issues[0]?.message,
-      );
-      return;
-    }
-
-    submitDraftAndFinish();
-  };
-
   return (
     <>
       <div className="mx-auto w-[90%] pt-4 md:pt-20">
@@ -77,23 +12,7 @@ const FormSection = () => {
         </p>
 
         <Kategori />
-
-        {category === "Daging" && <PertanyaanDaging />}
-        {category === "Sayur" && <PertanyaanSayur />}
-        {category === "Buah" && <PertanyaanBuah />}
-        {category === "Hasil Ternak" && <PertanyaanTernak />}
-
-        <div className="flex origin-right items-center justify-between">
-          <ButtonPrev
-            onClick={() => setStep(isQuestionOpen ? step - 2 : step - 1)}
-            href="/form-1"
-          >
-            Kembali
-          </ButtonPrev>
-          <ButtonNext onClick={handleSubmit} disabled={isSubmittingDraft}>
-            Lanjutkan
-          </ButtonNext>
-        </div>
+        <FormSectionInteractive />
       </div>
     </>
   );
