+++ src/shared/component/auth/FormInput.tsx
@@ -33,7 +33,7 @@ export const FormInput = ({
         value={value}
         onChange={onChange}
         placeholder={placeholder}
-        className={`peer bg-white py-3 text-[10px] ring-2 transition-all duration-200 placeholder:text-[10px] placeholder:font-medium md:py-5 md:text-[14px] md:placeholder:text-[14px] ${className} ${leftIcon ? "px-4 placeholder-shown:px-10" : "px-4"} ${rightIcon || isPassword ? "pr-10" : ""}`}
+        className={`peer bg-white py-3 text-[10px] transition-all duration-200 placeholder:text-[10px] placeholder:font-medium md:py-5 md:text-[14px] md:placeholder:text-[14px] ${className} ${leftIcon ? "px-4 placeholder-shown:px-10" : "px-4"} ${rightIcon || isPassword ? "pr-10" : ""}`}
       />
 
       {leftIcon && (
